+++ test/unit/builtins/common.py
@@ -3,6 +3,8 @@ from .. import AttrDict, FileTestCase, make_env, TestCase  # noqa: F401
 from bfg9000 import file_types
 from bfg9000.builtins import builtin
 from bfg9000.build_inputs import BuildInputs
+from bfg9000.options import option_list
+from bfg9000.packages import Package
 from bfg9000.path import Path, Root
 
 
@@ -21,3 +23,18 @@ class BuiltinTestCase(FileTestCase):
             builtin.BuildContext.PathEntry(build.bfgpath)
         )
         return build, context
+
+
+class MockPackage(Package):
+    def __init__(self, name, submodules=None, version=None, *, format,
+                 compile_options=None, link_options=None):
+        super().__init__(name, submodules, format=format)
+        self.version = version
+        self._compile_options = compile_options or option_list()
+        self._link_options = link_options or option_list()
+
+    def compile_options(self, compiler, *, raw=False):
+        return self._compile_options
+
+    def link_options(self, linker, *, raw=False):
+        return self._link_options