+++ test/unit/builtins/test_compile.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 from unittest import mock
 
-from .common import AttrDict, BuiltinTestCase
+from .common import AttrDict, BuiltinTestCase, MockPackage
 from .. import make_env
 
 from bfg9000 import file_types, options as opts
@@ -11,7 +11,6 @@ from bfg9000.builtins import (compile, link, packages, project,  # noqa: F401
                               regenerate)
 from bfg9000.environment import LibraryMode
 from bfg9000.iterutils import listify, unlistify
-from bfg9000.packages import CommonPackage
 from bfg9000.path import Path, Root
 from bfg9000.tools.msvc import MsvcBuilder
 
@@ -200,8 +199,8 @@ class TestObjectFile(CompileTest):
         pkg_incdir = opts.include_dir(file_types.HeaderDirectory(
             Path('/usr/include', Root.absolute)
         ))
-        pkg = CommonPackage('pkg', format=fmt,
-                            compile_options=opts.option_list(pkg_incdir))
+        pkg = MockPackage('pkg', format=fmt,
+                          compile_options=opts.option_list(pkg_incdir))
 
         result = self.context['object_file'](file='main.cpp',
                                              includes='include', packages=pkg)